feat: add rudimentary logo (ico + svg) [#46]
for the ico i basically took a screenshot and converted it to ico in gimp

turns out the svg is more important though as that is the one used for the browser tabs

for the svg i exported an alpha selection to a vector format in gimp and filled these groups via
inkscape. looks really bad up close but for the browser tab preview it looks pretty decent.

obviously can be replaced once we have a real logo :D